pyproject.toml
[tool.poetry]
name = "claim-agent"
version = "0.1.0"
description = "Agentic RAG Insurance Claim Processing System"
authors = ["Alireza"]
readme = "README.md"
packages = [{ include = "claim_agent", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"

# Core
pydantic = ">=2.0"
hydra-core = ">=1.3"
omegaconf = "*"
chromadb = ">=0.4"
sentence-transformers = "*"
pypdf2 = "*"
openai = ">=1.0"
pandas = "*"

# LangChain pipeline
langchain = ">=0.3"
langchain-openai = "*"
langchain-community = "*"
langchain-chroma = "*"
langgraph = ">=0.2"
duckduckgo-search = "*"

# Smolagents pipeline
smolagents = { extras = ["toolkit"], version = "*" }

# API
fastapi = ">=0.110"
uvicorn = { extras = ["standard"], version = "*" }

# Frontend
streamlit = ">=1.30"
requests = "*"

# Logging
loguru = "*"

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-asyncio = "*"
httpx = "*"
ruff = "*"
mypy = "*"
pre-commit = "*"


[tool.poetry.group.docs.dependencies]
mkdocs-material = "^9.7.3"
mkdocs = "^1.6.1"
mkdocs-mermaid2-plugin = "^1.2.3"
pymdown-extensions = "^10.21"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
